fix(ts): 修复pnpm build时，类型报错。需要添加 zod依赖
