+++ packages/sdk/package.json
@@ -2,6 +2,7 @@
   "name": "@opentiny/next-sdk",
   "version": "0.0.1-alpha.3",
   "description": "",
+  "type": "module",
   "module": "src/index.ts",
   "types": "src/index.ts",
   "publishConfig": {
@@ -26,6 +27,7 @@
   "dependencies": {
     "@modelcontextprotocol/sdk": "^1.15.1",
     "@opentiny/next": "^0.2.1",
+    "zod": "^3.25.10",
     "openai": "^5.9.0"
   },
   "devDependencies": {
