--- packages/sdk/src/client/index.ts
@@ -1,1 +1,0 @@
-export * from './createClient'
