--- packages/sdk/src/server/index.ts
@@ -1,1 +1,0 @@
-export * from './createServer'
